Refactored out ScriptFileWatcher from Scripting
