+++ src/main/java/org/openpnp/gui/MainFrame.java
@@ -92,6 +92,7 @@ import org.openpnp.gui.support.OSXAdapter;
 import org.openpnp.gui.support.RotationCellValue;
 import org.openpnp.model.Configuration;
 import org.openpnp.model.LengthUnit;
+import org.openpnp.scripting.ScriptFileWatcher;
 import org.pmw.tinylog.Logger;
 
 import com.jgoodies.forms.layout.ColumnSpec;
@@ -219,6 +220,8 @@ public class MainFrame extends JFrame {
 
     private ActionListener instructionsCancelActionListener;
     private ActionListener instructionsProceedActionListener;
+
+    private ScriptFileWatcher scriptFileWatcher;
 
     public MainFrame(Configuration configuration) {
         mainFrame = this;
@@ -654,7 +657,8 @@ public class MainFrame extends JFrame {
         while (!configurationLoaded) {
 	        try {
 	            configuration.load();
-	            configuration.getScripting().setMenu(mnScripts);
+	            scriptFileWatcher = new ScriptFileWatcher(configuration.getScripting());
+	            scriptFileWatcher.setMenu(mnScripts);
 	            
 	            if (Configuration.get().getMachine().getProperty("Welcome2_0_Dialog_Shown") == null) {
 	                Welcome2_0Dialog dialog = new Welcome2_0Dialog(this);
